CR with Shay/Lior

+++ signals.cpp
@@ -16,8 +16,10 @@ extern Shell *shell;
 
 void SignalHandler_ctrZ(int signal)
 {
+   int status;
+   waitpid(shell->fgJob.PID, &status, WNOHANG);
    cout << "smash: caught ctrl-Z" << endl;
-   if(shell->fgJob.PID != NOT_EXCIST)
+   if((shell->fgJob.PID != NOT_EXCIST) && WIFSTOPPED(status))
    {
       if(kill(shell->fgJob.PID, SIGSTOP) == 0)
       {
@@ -32,8 +34,10 @@ void SignalHandler_ctrZ(int signal)
 
 void SignalHandler_ctrC(int signal)
 {
+   int status;
+   waitpid(shell->fgJob.PID, &status, WNOHANG);
    cout << "smash: caught ctrl-C" << endl;
-   if(shell->fgJob.PID != NOT_EXCIST)
+   if((shell->fgJob.PID != NOT_EXCIST) && WIFSIGNALED(status))
    {
       if(kill(shell->fgJob.PID, SIGKILL) == 0)
       {
